+++ src/test/converter2/issues/gh2876.ts
@@ -1,0 +1,13 @@
+export function MyComponent(props: MyComponent.Props) {
+    return;
+}
+
+export namespace MyComponent {
+    export interface Props {
+        children?: {};
+    }
+}
+
+MyComponent.propTypes = {
+    children: {},
+};